move text output into renderer package
